requirements.txt
fastapi
uvicorn
requests
openai
pandas
pillow
python-dateutil
httpx
markdown
duckdb
scikit-learn
numpy
easyocr
faker
python-dotenv
gitpython
beautifulsoup4
sentence-transformers
scipy
